refactor: applied recommended go project structure
