+++ back/internal/models/rate.go
