+++ back/internal/repositories/base_repository.go
@@ -1,4 +1,4 @@
-package database
+package repositories
 
 import (
 	"context"
