+++ back/internal/services/rate_service.go
@@ -1,0 +1,16 @@
+package services
+
+import (
+	"github.com/ekosachev/movie-hub/internal/models"
+	"github.com/ekosachev/movie-hub/internal/repositories"
+)
+
+type RateService struct {
+	*BaseService[models.Rate]
+}
+
+func NewRateService(repo repositories.BaseRepository[models.Rate]) *RateService {
+	return &RateService{
+		BaseService: NewBaseService(repo),
+	}
+}
