+++ back/internal/services/reaction_service.go
@@ -1,0 +1,16 @@
+package services
+
+import (
+	"github.com/ekosachev/movie-hub/internal/models"
+	"github.com/ekosachev/movie-hub/internal/repositories"
+)
+
+type ReactionService struct {
+	*BaseService[models.Reaction]
+}
+
+func NewReactionService(repo repositories.BaseRepository[models.Reaction]) *ReactionService {
+	return &ReactionService{
+		BaseService: NewBaseService(repo),
+	}
+}
