+++ back/internal/services/role_repository.go
@@ -1,0 +1,16 @@
+package services
+
+import (
+	"github.com/ekosachev/movie-hub/internal/models"
+	"github.com/ekosachev/movie-hub/internal/repositories"
+)
+
+type RoleService struct {
+	*BaseService[models.Role]
+}
+
+func NewRoleService(repo repositories.BaseRepository[models.Role]) *RoleService {
+	return &RoleService{
+		BaseService: NewBaseService(repo),
+	}
+}
