--- back/services/movie_cast_service.go
@@ -1,16 +1,0 @@
-package services
-
-import (
-	"github.com/ekosachev/movie-hub/database"
-	"github.com/ekosachev/movie-hub/database/models"
-)
-
-type MoveCastService struct {
-	*BaseService[models.MovieCast]
-}
-
-func NewMovieCastService(repo database.BaseRepository[models.MovieCast]) *MoveCastService {
-	return &MoveCastService{
-		BaseService: NewBaseService(repo),
-	}
-}
