--- back/services/movie_service.go
@@ -1,16 +1,0 @@
-package services
-
-import (
-	"github.com/ekosachev/movie-hub/database"
-	"github.com/ekosachev/movie-hub/database/models"
-)
-
-type MovieService struct {
-	*BaseService[models.Movie]
-}
-
-func NewMovieService(repo database.BaseRepository[models.Movie]) *MovieService {
-	return &MovieService{
-		BaseService: NewBaseService(repo),
-	}
-}
