--- back/services/role_repository.go
@@ -1,16 +1,0 @@
-package services
-
-import (
-	"github.com/ekosachev/movie-hub/database"
-	"github.com/ekosachev/movie-hub/database/models"
-)
-
-type RoleService struct {
-	*BaseService[models.Role]
-}
-
-func NewRoleService(repo database.BaseRepository[models.Role]) *RoleService {
-	return &RoleService{
-		BaseService: NewBaseService(repo),
-	}
-}
